+++ lib/extensions/invitation/ecto/schema.ex
@@ -1,9 +1,13 @@
 defmodule PowInvitation.Ecto.Schema do
-  @moduledoc false
+  @moduledoc """
+  Handles the invitation schema for user.
+  """
+
   use Pow.Extension.Ecto.Schema.Base
   alias Ecto.Changeset
   alias Pow.UUID
 
+  @doc false
   @impl true
   def attrs(_config) do
     [
@@ -12,6 +16,7 @@ defmodule PowInvitation.Ecto.Schema do
     ]
   end
 
+  @doc false
   @impl true
   def assocs(_config) do
     [
@@ -20,6 +25,7 @@ defmodule PowInvitation.Ecto.Schema do
     ]
   end
 
+  @doc false
   @impl true
   def indexes(_config) do
     [{:invitation_token, true}]
@@ -36,6 +42,18 @@ defmodule PowInvitation.Ecto.Schema do
     end
   end
 
+  @doc """
+  Invites user.
+
+  It's important to note that this changeset should not include the changeset
+  method in the user schema module if `PowEmailConfirmation` has been enabled.
+  This is because the e-mail is assumed confirmed already if the user can
+  accept the invitation.
+
+  A unique `:invitation_token` will be generated, and `invited_by` association
+  will be set. Only the user id will be set, and the persisted user won't have
+  any password for authentication.
+  """
   @spec invite_changeset(Ecto.Schema.t() | Changeset.t(), Ecto.Schema.t(), map()) :: Changeset.t()
   def invite_changeset(%Changeset{data: user} = changeset, invited_by, attrs) do
     changeset
@@ -61,6 +79,13 @@ defmodule PowInvitation.Ecto.Schema do
     Changeset.assoc_constraint(%{changeset | data: data}, :invited_by)
   end
 
+  @doc """
+  Accepts an invitation.
+
+  The changeset method in user schema module is called, and
+  `:invitation_accepted_at` will be updated. The password can be set, and the
+  user id updated.
+  """
   @spec accept_invitation_changeset(Ecto.Schema.t() | Changeset.t(), map()) :: Changeset.t()
   def accept_invitation_changeset(%Changeset{data: %user_mod{}} = changeset, attrs) do
     accepted_at = Pow.Ecto.Schema.__timestamp_for__(user_mod, :invitation_accepted_at)